moved tests, added dummy ComputeECDHKey
