+++ lib/ocrypto/rsa_key_pair_test.go
@@ -1,8 +1,6 @@
 package ocrypto
 
 import (
-	"crypto/sha256"
-	"github.com/stretchr/testify/require"
 	"testing"
 )
 
@@ -50,46 +48,5 @@ func TestRSAKeyPair(t *testing.T) {
 	_, err = emptyRSAKeyPair.KeySize()
 	if err == nil {
 		t.Fatal("RsaKeyPair.keySize() fail to return error")
-	}
-}
-
-func TestNanoTDFRewrapKeyGenerate(t *testing.T) {
-	kasECKeyPair, err := NewECKeyPair(ECCModeSecp256r1)
-	require.NoError(t, err, "fail on NewECKeyPair")
-
-	kasPubKeyAsPem, err := kasECKeyPair.PublicKeyInPemFormat()
-	require.NoError(t, err, "fail to generate ec public key in pem format")
-
-	kasPrivateKeyAsPem, err := kasECKeyPair.PrivateKeyInPemFormat()
-	require.NoError(t, err, "fail to generate ec private key in pem format")
-
-	sdkECKeyPair, err := NewECKeyPair(ECCModeSecp256r1)
-	require.NoError(t, err, "fail on NewECKeyPair")
-
-	sdkPubKeyAsPem, err := sdkECKeyPair.PublicKeyInPemFormat()
-	require.NoError(t, err, "fail to generate ec public key in pem format")
-
-	sdkPrivateKeyAsPem, err := sdkECKeyPair.PrivateKeyInPemFormat()
-	require.NoError(t, err, "fail to generate ec private key in pem format")
-
-	kasECDHKey, err := ComputeECDHKey(kasPrivateKeyAsPem, sdkPubKeyAsPem)
-	require.NoError(t, err, "fail to calculate ecdh key")
-
-	// slat
-	digest := sha256.New()
-	digest.Write([]byte("L1L"))
-
-	kasSymmetricKey, err := CalculateHKDF(digest.Sum(nil), kasECDHKey, 32)
-	require.NoError(t, err, "fail to calculate HKDF key")
-
-	sdkECDHKey, err := ComputeECDHKey(sdkPrivateKeyAsPem, kasPubKeyAsPem)
-	require.NoError(t, err, "fail to calculate ecdh key")
-
-	sdkSymmetricKey, err := CalculateHKDF(digest.Sum(nil), sdkECDHKey, 32)
-	require.NoError(t, err, "fail to calculate HKDF key")
-
-	if string(kasSymmetricKey) != string(sdkSymmetricKey) {
-		t.Fatalf("symmetric keys on both kas and sdk should be same kas:%s sdk:%s",
-			string(kasSymmetricKey), string(sdkSymmetricKey))
 	}
 }
